clean-up for texture view preview
