+++ camera/settings.gradle
@@ -1,2 +1,2 @@
 include ':basic'
-include ':texture-preview'
+include ':texture-view'
