--- camera/texture-preview/src/main/res/values/strings.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <string name="app_name">CameraTextureView</string>
-    <string name="action_settings">Settings</string>
-
-</resources>
