+++ camera/texture-view/build.gradle
@@ -1,8 +1,8 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 24
-    buildToolsVersion "25.0.0"
+    compileSdkVersion 25
+    buildToolsVersion "25.0.3"
 
     defaultConfig {
         applicationId "com.sample.textureview"
@@ -36,5 +36,5 @@ android {
 }
 
 dependencies {
-    compile 'com.android.support:support-compat:24.2.0'
+    compile 'com.android.support:support-compat:25.3.1'
 }
