+++ camera/texture-view/src/main/cpp/CMakeLists.txt
