+++ camera/texture-view/src/main/cpp/android_main.cpp
@@ -23,15 +23,15 @@
  *         Google Pixel and Nexus 6 phones
  */
 #include <jni.h>
+#include <utils/native_debug.h>
 #include "camera_manager.h"
 #include "camera_engine.h"
-
 
 /**
  * Application object:
  *   the top level camera application object, maintained by native code only
  */
-CameraEngine *pEngineObj = nullptr;
+CameraAppEngine *pEngineObj = nullptr;
 
 /**
  * createCamera() Create application instance and NDK camera object
@@ -47,16 +47,10 @@ CameraEngine *pEngineObj = nullptr;
  * @return application object instance ( not used in this sample )
  */
 extern "C" JNIEXPORT jlong JNICALL
-Java_com_sample_textureview_ViewActivity_createCamera(
-        JNIEnv *env,
-        jobject instance,
-        jint width,
-        jint height,
-        jint rotation) {
-
-  pEngineObj = new CameraEngine(env, width, height, rotation);
-  pEngineObj->CreateCamera();
-
+Java_com_sample_textureview_ViewActivity_createCamera(JNIEnv *env,
+                                                      jobject instance,
+                                                      jint width, jint height) {
+  pEngineObj = new CameraAppEngine(env, instance, width, height);
   return reinterpret_cast<jlong>(pEngineObj);
 }
 
@@ -66,13 +60,18 @@ Java_com_sample_textureview_ViewActivity_createCamera(
  *   triggers native camera object be released
  */
 extern "C" JNIEXPORT void JNICALL
-Java_com_sample_textureview_ViewActivity_deleteCamera(
-    JNIEnv *env, jobject instance) {
-  if(!pEngineObj) {
+Java_com_sample_textureview_ViewActivity_deleteCamera(JNIEnv *env,
+                                                      jobject instance,
+                                                      jlong ndkCameraObj) {
+  if (!pEngineObj || !ndkCameraObj) {
     return;
   }
+  CameraAppEngine *pApp = reinterpret_cast<CameraAppEngine *>(ndkCameraObj);
+  ASSERT(pApp == pEngineObj, "NdkCamera Obj mismatch");
 
-  delete pEngineObj;
+  delete pApp;
+
+  // also reset the private global object
   pEngineObj = nullptr;
 }
 
@@ -88,19 +87,17 @@ Java_com_sample_textureview_ViewActivity_deleteCamera(
  *      on display device
  */
 extern "C" JNIEXPORT jobject JNICALL
-Java_com_sample_textureview_ViewActivity_getMinimumCompatiblePreviewSize(JNIEnv *env,
-                                                                          jobject instance,
-                                                                          jlong nativeObj) {
-
-  if (!nativeObj) {
+Java_com_sample_textureview_ViewActivity_getMinimumCompatiblePreviewSize(
+    JNIEnv *env, jobject instance, jlong ndkCameraObj) {
+  if (!ndkCameraObj) {
     return nullptr;
   }
-  CameraEngine *pApp = reinterpret_cast<CameraEngine*>(nativeObj);
+  CameraAppEngine *pApp = reinterpret_cast<CameraAppEngine *>(ndkCameraObj);
   jclass cls = env->FindClass("android/util/Size");
-  jobject previewSize = env->NewObject(cls,
-                                   env->GetMethodID(cls, "<init>", "(II)V"),
-                                   pApp->GetCompatibleCameraRes().width,
-                                   pApp->GetCompatibleCameraRes().height);
+  jobject previewSize =
+      env->NewObject(cls, env->GetMethodID(cls, "<init>", "(II)V"),
+                     pApp->GetCompatibleCameraRes().width,
+                     pApp->GetCompatibleCameraRes().height);
   return previewSize;
 }
 
@@ -111,8 +108,10 @@ Java_com_sample_textureview_ViewActivity_getMinimumCompatiblePreviewSize(JNIEnv 
  */
 extern "C" JNIEXPORT jint JNICALL
 Java_com_sample_textureview_ViewActivity_getCameraSensorOrientation(
-    JNIEnv *env, jobject instance) {
-  return pEngineObj->GetCameraSensorOrientation(ACAMERA_LENS_FACING_BACK);
+    JNIEnv *env, jobject instance, jlong ndkCameraObj) {
+  ASSERT(ndkCameraObj, "NativeObject should not be null Pointer");
+  CameraAppEngine *pApp = reinterpret_cast<CameraAppEngine *>(ndkCameraObj);
+  return pApp->GetCameraSensorOrientation(ACAMERA_LENS_FACING_BACK);
 }
 
 /**
@@ -123,9 +122,12 @@ Java_com_sample_textureview_ViewActivity_getCameraSensorOrientation(
  */
 extern "C" JNIEXPORT void JNICALL
 Java_com_sample_textureview_ViewActivity_onPreviewSurfaceCreated(
-    JNIEnv *env, jobject instance, jobject surface) {
-   pEngineObj->CreateCameraSession(surface);
-   pEngineObj->StartPreview(true);
+    JNIEnv *env, jobject instance, jlong ndkCameraObj, jobject surface) {
+  ASSERT(ndkCameraObj && (jlong)pEngineObj == ndkCameraObj,
+         "NativeObject should not be null Pointer");
+  CameraAppEngine *pApp = reinterpret_cast<CameraAppEngine *>(ndkCameraObj);
+  pApp->CreateCameraSession(surface);
+  pApp->StartPreview(true);
 }
 
 /**
@@ -133,12 +135,29 @@ Java_com_sample_textureview_ViewActivity_onPreviewSurfaceCreated(
  *   Notification to native camera that java TextureView is destroyed
  *   Native camera would:
  *      * stop preview
- *      * delete appliation object
  */
 extern "C" JNIEXPORT void JNICALL
 Java_com_sample_textureview_ViewActivity_onPreviewSurfaceDestroyed(
-    JNIEnv *env, jobject instance, jobject surface) {
-  pEngineObj->StartPreview(false);
-  delete pEngineObj;
-  pEngineObj = nullptr;
+    JNIEnv *env, jobject instance, jlong ndkCameraObj, jobject surface) {
+  CameraAppEngine *pApp = reinterpret_cast<CameraAppEngine *>(ndkCameraObj);
+  ASSERT(ndkCameraObj && pEngineObj == pApp,
+         "NativeObject should not be null Pointer");
+  jclass cls = env->FindClass("android/view/Surface");
+  jmethodID toString =
+      env->GetMethodID(cls, "toString", "()Ljava/lang/String;");
+
+  jstring destroyObjStr =
+      reinterpret_cast<jstring>(env->CallObjectMethod(surface, toString));
+  const char *destroyObjName = env->GetStringUTFChars(destroyObjStr, nullptr);
+
+  jstring appObjStr = reinterpret_cast<jstring>(
+      env->CallObjectMethod(pApp->GetSurfaceObject(), toString));
+  const char *appObjName = env->GetStringUTFChars(appObjStr, nullptr);
+
+  ASSERT(!strcmp(destroyObjName, appObjName), "object Name MisMatch");
+
+  env->ReleaseStringUTFChars(destroyObjStr, destroyObjName);
+  env->ReleaseStringUTFChars(appObjStr, appObjName);
+
+  pApp->StartPreview(false);
 }
