+++ camera/texture-view/src/main/cpp/camera_listeners.cpp
