+++ camera/texture-view/src/main/res/drawable-hdpi/ic_launcher.png
