+++ camera/texture-view/src/main/res/drawable-mdpi/ic_launcher.png
