+++ camera/texture-view/src/main/res/drawable-xhdpi/ic_launcher.png
