+++ camera/texture-view/src/main/res/drawable-xxhdpi/ic_launcher.png
